requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
PyMySQL==1.1.1
cryptography==43.0.3
alembic==1.14.0
pydantic>=2.10.3
pydantic-settings>=2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
apscheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1
